+++ scripts/winesapos-tests.sh
@@ -532,6 +532,7 @@ for i in \
 done
 
 for i in \
+  ${WINESAPOS_INSTALL_DIR}/home/winesap/Desktop/appimagelauncher.desktop \
   ${WINESAPOS_INSTALL_DIR}/home/winesap/Desktop/blueman-manager.desktop \
   ${WINESAPOS_INSTALL_DIR}/home/winesap/Desktop/com.usebottles.bottles.desktop \
   ${WINESAPOS_INSTALL_DIR}/home/winesap/Desktop/org.gnome.Cheese.desktop \
@@ -750,15 +751,17 @@ if [[ "${WINESAPOS_DISTRO}" != "manjaro" ]]; then
 fi
 echo "Checking that all the packages from the AUR have been installed by yay done."
 
-echo 'Testing that the "pamac" package manager is installed...'
+echo "Testing that the GUI package managers Pamac and AppImageLauncher are installed..."
 if [[ "${WINESAPOS_DISTRO}" == "manjaro" ]]; then
     pacman_search_loop \
       pamac-gtk pamac-cli \
       libpamac-flatpak-plugin \
-      libpamac-snap-plugin
+      libpamac-snap-plugin \
+      appimagelauncher
 else
     pacman_search_loop \
-      pamac-all
+      pamac-all \
+      appimagelauncher
 fi
 
 echo "Checking that all of the Pamac plugins are enabled..."
@@ -772,7 +775,7 @@ for i in EnableAUR CheckAURUpdates EnableFlatpak CheckFlatpakUpdates EnableSnap;
     fi
 done
 echo "Checking that all of the Pamac plugins are enabled complete."
-echo "Testing that the "pamac" package manager is installed complete."
+echo "Testing that the GUI package managers Pamac and AppImageLauncher are installed complete."
 
 echo 'Testing that the PipeWire audio library is installed...'
 echo "Checking that PipeWire packages are installed..."